Added final project model
Final project model is the code from the step-by-step directions in the course, in addition to the model extensions and instance data from the course dataset. This file should be used by learners after they have completed the course and wish to build out the model as their final project.